Project objects:
The project consists of two parts, software, and hardware. In the software part, the mobile application of the system is carried out in the AppInventor environment. In the hardware part, the necessary electronic circuit is created. There is a microcontroller in the electronic circuit part, a Bluetooth module, and other electronic circuit elements to transfer the information from the user to the microcontroller. The printed circuit design of the circuit was carried out using the EasyEda interface.